requirements.txt
pandas
numpy
matplotlib
scikit-learn
openpyxl
seaborn
streamlit
ipykernel
plotly.express
ast